move isyntax fixture to conftest
