+++ tests/test_isyntax_wsi.py
@@ -1,25 +1,10 @@
-import urllib.request
 from pathlib import Path
 from tempfile import TemporaryDirectory
 
-import pytest
 from isyntax import ISyntax
 from tifffile import TiffFile
 
 from panimg.contrib.wsi_isyntax_to_tiff.isyntax_to_tiff import isyntax_to_tiff
-from tests import RESOURCE_PATH
-
-
-@pytest.fixture(scope="session", autouse=True)
-def downloaded_isyntax_image():
-    url = "https://zenodo.org/record/5037046/files/testslide.isyntax"
-    image_path = RESOURCE_PATH / "isyntax_wsi" / "testslide.isyntax"
-    image_path.parent.mkdir()
-
-    if not image_path.exists():
-        urllib.request.urlretrieve(url, image_path)
-
-    return image_path
 
 
 def test_isyntax_to_tiff(downloaded_isyntax_image):
